buyerservice/requirements.txt
fastapi
uvicorn
motor
python-dotenv
PyJWT
razorpay
